requirements.txt
keras-ocr
googletrans==4.0.0-rc1
opencv-python
Pillow
instaloader
instabot